Refactor model tests.
