+++ tests/test_model.py
@@ -1,8 +1,7 @@
 import os
 import tempfile
-import time
 import shutil
-from typing import Text, Optional
+from typing import Text
 
 import pytest
 from _pytest.tmpdir import TempdirFactory
@@ -10,16 +9,9 @@ from _pytest.tmpdir import TempdirFactory
 import rasa
 import rasa.core
 import rasa.nlu
-from rasa.importers.rasa import RasaFileImporter
-from rasa.constants import DEFAULT_CONFIG_PATH, DEFAULT_DATA_PATH
 from rasa.core.domain import Domain
 from rasa.model import (
-    FINGERPRINT_CONFIG_KEY,
-    FINGERPRINT_DOMAIN_KEY,
     FINGERPRINT_FILE_PATH,
-    FINGERPRINT_NLU_DATA_KEY,
-    FINGERPRINT_RASA_VERSION_KEY,
-    FINGERPRINT_STORIES_KEY,
     FINGERPRINT_TRAINED_AT_KEY,
     core_fingerprint_changed,
     create_package_rasa,
@@ -30,10 +22,33 @@ from rasa.model import (
     nlu_fingerprint_changed,
     Fingerprint,
     should_retrain,
-    FINGERPRINT_CONFIG_CORE_KEY,
-    FINGERPRINT_CONFIG_NLU_KEY,
+    unpack_model,
+    package_model,
 )
 from rasa.exceptions import ModelNotFound
+from tests.utilities import _fingerprint, _project_files
+
+
+def test_unpack_model(trained_model):
+    unpacked_model_file = unpack_model(trained_model)
+
+    assert os.path.isdir(unpacked_model_file)
+    assert os.path.exists(unpacked_model_file)
+    assert os.path.exists(trained_model)
+
+
+def test_package_model(trained_model, tmpdir_factory):
+    unpacked_model_file = unpack_model(trained_model)
+
+    output_path = tmpdir_factory.mktemp("model").strpath
+
+    packed_model = package_model(
+        _fingerprint(), output_path, unpacked_model_file, "model-file-name"
+    )
+
+    assert os.path.exists(packed_model)
+    assert not os.path.exists(unpacked_model_file)
+    assert os.path.basename(packed_model) == "model-file-name.tar.gz"
 
 
 def test_get_latest_model(trained_model):
@@ -90,29 +105,6 @@ def test_get_model_from_directory_nlu_only(trained_model):
     assert unpacked_nlu
 
 
-def _fingerprint(
-    config: Optional[Text] = None,
-    config_nlu: Optional[Text] = None,
-    config_core: Optional[Text] = None,
-    domain: Optional[int] = None,
-    rasa_version: Text = "1.0",
-    stories: Optional[int] = None,
-    nlu: Optional[int] = None,
-):
-    return {
-        FINGERPRINT_CONFIG_KEY: config if config is not None else ["test"],
-        FINGERPRINT_CONFIG_CORE_KEY: config_core
-        if config_core is not None
-        else ["test"],
-        FINGERPRINT_CONFIG_NLU_KEY: config_nlu if config_nlu is not None else ["test"],
-        FINGERPRINT_DOMAIN_KEY: domain if domain is not None else ["test"],
-        FINGERPRINT_TRAINED_AT_KEY: time.time(),
-        FINGERPRINT_RASA_VERSION_KEY: rasa_version,
-        FINGERPRINT_STORIES_KEY: stories if stories is not None else ["test"],
-        FINGERPRINT_NLU_DATA_KEY: nlu if nlu is not None else ["test"],
-    }
-
-
 def test_persist_and_load_fingerprint():
     from rasa.model import persist_fingerprint, fingerprint_from_path
 
@@ -153,24 +145,6 @@ def test_core_fingerprint_changed(fingerprint2):
 def test_nlu_fingerprint_changed(fingerprint2):
     fingerprint1 = _fingerprint()
     assert nlu_fingerprint_changed(fingerprint1, fingerprint2)
-
-
-def _project_files(
-    project,
-    config_file=DEFAULT_CONFIG_PATH,
-    domain="domain.yml",
-    training_files=DEFAULT_DATA_PATH,
-):
-    paths = {
-        "config_file": config_file,
-        "domain_path": domain,
-        "training_data_paths": training_files,
-    }
-
-    paths = {k: v if v is None else os.path.join(project, v) for k, v in paths.items()}
-    paths["training_data_paths"] = [paths["training_data_paths"]]
-
-    return RasaFileImporter(**paths)
 
 
 async def test_create_fingerprint_from_paths(project):
